content/events/test/test_bug563329.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=563329
-->
<head>
  <title>Test for Bug 563329</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=563329">Mozilla Bug 563329</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript;version=1.7">

/** Test for Bug 563329 **/
/*  ui.click_hold_context_menus preference */

var target = null;
var tests = getTests();
var currentTest = null;

function getTests() {
  let tests = [
    { "func": function() { setTimeout(doCheckContextMenu, 100)}, "message": "Context menu should has fired"},
    { "func": function() { setTimeout(doCheckDuration, 100)}, "message": "Context menu should has fired with delay"},
    { "func": function() { setTimeout(finishTest, 100)}, "message": "" }
  ];

  let i = 0;
  while (i < tests.length)
    yield tests[i++];
}

function doTest() {
  // Enable context menus
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  var prefs = Components.classes["@mozilla.org/preferences-service;1"].
              getService(Components.interfaces.nsIPrefBranch);
  prefs.setBoolPref("ui.click_hold_context_menus", true);

  target = document.getElementById("testTarget");

  document.documentElement.addEventListener("contextmenu", function() {
    SimpleTest.ok(true, currentTest.message);
    synthesizeMouse(target, 0, 0, {type: "mouseup"});
    SimpleTest.executeSoon(function() {
      currentTest = tests.next();
      currentTest.func();
    });
  }, false);

  SimpleTest.executeSoon(function() {
    currentTest = tests.next();
    currentTest.func();
  });
}

function doCheckContextMenu() {
  synthesizeMouse(target, 0, 0, {type: "mousedown"});
}

function doCheckDuration() {
  var duration = 50;

  // Change click hold delay
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  var prefs = Components.classes["@mozilla.org/preferences-service;1"].
              getService(Components.interfaces.nsIPrefBranch);
  prefs.setIntPref("ui.click_hold_context_menus.delay", duration);

  synthesizeMouse(target, 0, 0, {type: "mousedown"});
}

function finishTest() {
  synthesizeKey("VK_ESCAPE", {}, window);
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  var prefs = Components.classes["@mozilla.org/preferences-service;1"].
              getService(Components.interfaces.nsIPrefBranch);
  try {
    prefs.clearUserPref("ui.click_hold_context_menus");
  }
  catch(e) {}

  try {
    prefs.clearUserPref("ui.click_hold_context_menus.delay");
  }
  catch(e) {}

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(doTest);

</script>
</pre>
<span id="testTarget" style="border: 1px solid black;">testTarget</span>
</body>
</html>
